requirements.txt
requests==2.6.2
prettytable==0.7.2
PyYAML==3.11
tailer==0.3
arrow==0.6.0
pytest==2.7.0
setuptools==18.4
